(cmake): PoC new way of combining binaries
This method is based off of picowota (https://github.com/usedbytes/picowota) as well as the pico sdk itself.

With this method we can create all of the expected output files (.elf, .hex, .uf2) as well as register it as an executable in cmake.

This commit is just a proof of concept, in later commits this will be extended to have the functionality described above.