(linker|cmake): Place the flashloader inside the main binary
In this commit we switch from placing the application in the flashloader to placing the flashloader in the application.
This makes it easier to debug the application as debug symbols are already in the correct location, at the expense of losing easy debugging for the flashloader.
Prior to this commit it was difficult to debug the application with the flashloader as the debug symbols did not line up.