(flashloader): Add required tusb dfu callbacks for device to successfully enumerate
With this commit the device is sucessfully enumerated and now presents a DFU interface. However, on Windows no driver is given to this interface so it cannot be claimed by WebUSB.

In Linux/macOS, this is not a problem.
To fix this, we must modify the Microsoft OS 2.0 descriptors to add an additional function subset header for the new interface.

For more information on the function subset header please visit: https://github.com/newaetech/naeusb/blob/main/wcid.md#function-subset-header