(core/tusb-conf): Add define scoped MS OS 2.0 values for DFU interface.
NOTE: At the moment, this change breaks the USB enumeration. I am curious to see if it is because of the size of "desc_ms_os_20".

Regardless, running this commit results in the pico panicking with the tusb error "ep 0 out was already available".

I will set up a example version of this so see if it can be duplicated outside of this repo, and if so, submit a bug to TinyUSB.