(flashloader/combined): Create sectioned hex files.
The stock .hex file contains two sections, the flashloader, starting at 0x10000000, and the application, currently starting at 0x10001000.

At present only the application is updated when using the DFU mode, so we need a way to only write the application to the flash.

We could hardcode offsets in the firmware but that would make it difficult when the offset is changed.

Instead, we are making seperate .hex files for each of the two sections.

We can then use micro:bit's universal hex format to merge them together in a way where we can clearly identify when each section starts without relying on fixed addresses.